Fix for N/A values on interface Error Rate graph
By default we were using the "ifconfig -a" output and attempting to
parse all of the per-interface datapoints from there. It turns out that
ifconfigs output is subtly different between Linux distributions and
versions, which makes it hard to parse successfully.

This change first attempts to parse the output of "ip -s -o link"
instead because it's far more stable, and more likely to exist on newer
Linux releases anyway.

Fixes ZEN-24388.